services/spnet-gram-backend/requirements.txt
requests>=2.31.0
google-auth>=2.0.0
